split GroupComponent, add edit modes
